+++ api/src/routes/coursesRoutes.js
@@ -10,6 +10,9 @@ const createCourse = require("../controllers/courses/createCourse");
 const updateCourse = require("../controllers/courses/updateCourse");
 
 const deleteCourse = require("../controllers/courses/deleteCourse");
+const filterOrdinances = require("../controllers/courses/filtersOrdiances");
+
+
 
 const router = Router();
 router.use(express.json());
@@ -27,5 +30,7 @@ router.put("/update/:id", updateCourse);
 
 // ----> DELETE <----
 router.post("/delete/:id", deleteCourse);
+router.post("/filters", filterOrdinances);
+
 
 module.exports = router;
