+++ client/src/redux/actions/filterByStars.js
@@ -1,11 +1,11 @@
 import axios from "axios";
 
 
-export function filterByStars(stars){
+export function filterByStars(stars, filtered){
 
   return async function(dispatch){
       try {
-        var json = await axios(`/courses?scoreToFilter=${stars?stars:""}`);
+        var json = await axios.post(`/courses?scoreToFilter=${stars?stars:""}`, filtered);
         return dispatch({
           type: "FILTER_BY_STARS",
           payload: json.data
