+++ client/src/redux/actions/filterRangeByPrice.js
@@ -1,10 +1,11 @@
 import axios from "axios";
 
 export function filterRangeByPrice(priceToFilter) {
+
   return async function (dispatch) {
     try {
-      let json = await axios(
-        `/courses?priceToFilter=${priceToFilter ? priceToFilter : ""}`
+      let json = await axios.post(
+        `/courses/filters?priceToFilter=${priceToFilter ? priceToFilter : ""}`
       );
       return dispatch({
         type: "FILTER_BY_RANGE_PRICE",
