+++ client/src/redux/actions/getAllCourses.js
@@ -6,7 +6,7 @@ export function getAllCourses() {
       var json = await axios(`/courses`);
       return dispatch({
         type: "GET_ALL_COURSES",
-        payload: { allCourses: json.data, name: "Filter By" },
+        payload: { allCourses: json.data, name2: "Filter By" },
       });
     } catch (error) {
       console.log("error", error);
