+++ client/src/redux/actions/getCourseByName.js
@@ -3,7 +3,7 @@ import axios from "axios";
 export function getCourseByName(name) {
   return async function (dispatch) {
     try {
-      var json = await axios(`/courses/search?name=${name ? name : ""}`);
+      var json = await axios.post(`/courses/search?name=${name ? name : ""}`);
       return dispatch({
         type: "GET_COURSE_BY_NAME",
         payload: json.data,
